// Client Server/client/src/model/Dependency.ts
/**
 * @author Michaluk David <49762@etu.he2b.be>
 * @author El Amouri Meihdi <49262@etu.he2b.be>
 * @author Rossitto Nicolas <49282@etu.he2b.be>
 * @author Azoud Ismael <42394@etu.he2b.be>
 * @licence GPL-3.0 <https://www.gnu.org/licenses/>
 */

/**
 * a package (or dependency) definition
 */
export default interface Dependency {
    name: string;
    version: string;
}